requirements.txt
amqp==2.3.2
argh==0.26.2
astroid==2.1.0
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
cloudpickle==0.6.1
cycler==0.10.0
Cython==0.29.2
dask==1.0.0
decorator==4.3.0
Django==2.1.4
django-dotenv==1.4.2
facebook-sdk==3.1.0
gunicorn==19.9.0
idna==2.8
isort==4.3.4
kiwisolver==1.0.1
kombu==4.2.1
lazy-object-proxy==1.3.1
matplotlib==3.0.2
mccabe==0.6.1
networkx==2.2
numpy==1.15.4
pathtools==0.1.2
Pillow==5.3.0
psycopg2-binary==2.7.6.1
pyclustering==0.8.2
pylint==2.2.2
pylint-django==2.0.5
pylint-plugin-utils==0.4
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
PyWavelets==1.0.1
PyYAML==3.13
redis==2.10.6
requests==2.21.0
scikit-image==0.14.1
scikit-learn==0.20.1
scipy==1.1.0
six==1.12.0
toolz==0.9.0
typed-ast==1.3.1
urllib3==1.24.1
vine==1.1.4
watchdog==0.9.0
wrapt==1.11.1
